requirements.txt
altgraph==0.17.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
msedge-selenium-tools==3.141.4
outcome==1.2.0
pefile==2023.2.7
Pillow==10.0.0
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PySocks==1.7.1
pystray==0.19.4
pywin32-ctypes==0.2.2
requests==2.31.0
schedule==1.2.0
selenium==3.141.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.1
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0
